+++ src/main/Application/ForReceipts.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.text.Font?>
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Application.ControllerForReceipts">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Application.ForReceiptsController">
    <children>
       <Label fx:id="ShowReceipts" layoutX="252.0" layoutY="102.0" />
       <Button fx:id="cancelButton" onAction = "#goToHome" cancelButton="true" layoutX="28.0" layoutY="21.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="30.0" text="&lt;">
